+++ tests/compiler/test_basics.py
@@ -1,0 +1,97 @@
+"""
+    Copyright 2018 Inmanta
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    Contact: code@inmanta.com
+"""
+
+from io import StringIO
+from itertools import groupby
+import os
+import re
+import shutil
+import sys
+import tempfile
+import unittest
+
+import pytest
+
+from inmanta import config
+from inmanta.ast import AttributeException, IndexException
+from inmanta.ast import MultiException
+from inmanta.ast import NotFoundException, TypingException
+from inmanta.ast import RuntimeException, DuplicateException, TypeNotFoundException, ModuleNotFoundException, \
+    OptionalValueException
+import inmanta.compiler as compiler
+from inmanta.execute.proxy import UnsetException
+from inmanta.execute.util import Unknown, NoneValue
+from inmanta.export import DependencyCycleException
+from inmanta.module import Project
+from inmanta.parser import ParserException
+from utils import assert_graph
+
+
+def test_str_on_instance_pos(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+import std
+
+entity Hg:
+end
+
+Hg.hosts [0:] -- std::Host
+
+implement Hg using std::none
+
+hg = Hg()
+
+for i in [1,2,3]:
+ hg.hosts = std::Host(name="Test{{i}}", os=std::unix)
+end
+
+
+for i in hg.hosts:
+    std::ConfigFile(host=i, path="/fx", content="")
+end
+""")
+    (types, _) = compiler.do_compile()
+    files = types["std::File"].get_all_instances()
+    assert len(files) == 3
+
+
+def test_str_on_instance_neg(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+import std
+
+entity Hg:
+end
+
+Hg.hosts [0:] -- std::Host
+
+implement Hg using std::none
+
+hg = Hg()
+
+for i in [1,2,3]:
+ hg.hosts = std::Host(name="Test", os=std::unix)
+end
+
+
+for i in hg.hosts:
+    std::ConfigFile(host=i, path="/fx", content="")
+end
+""")
+    (types, _) = compiler.do_compile()
+    files = types["std::File"].get_all_instances()
+    assert len(files) == 1
+
