+++ tests/compiler/test_dict.py
@@ -1,0 +1,121 @@
+"""
+    Copyright 2018 Inmanta
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    Contact: code@inmanta.com
+"""
+
+from io import StringIO
+from itertools import groupby
+import os
+import re
+import shutil
+import sys
+import tempfile
+import unittest
+
+import pytest
+
+from inmanta import config
+from inmanta.ast import AttributeException, IndexException
+from inmanta.ast import MultiException
+from inmanta.ast import NotFoundException, TypingException
+from inmanta.ast import RuntimeException, DuplicateException, TypeNotFoundException, ModuleNotFoundException, \
+    OptionalValueException
+import inmanta.compiler as compiler
+from inmanta.execute.proxy import UnsetException
+from inmanta.execute.util import Unknown, NoneValue
+from inmanta.export import DependencyCycleException
+from inmanta.module import Project
+from inmanta.parser import ParserException
+from utils import assert_graph
+
+
+def test_dict(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+a = "a"
+b = { "a" : a, "b" : "b", "c" : 3}
+""")
+
+    (_, root) = compiler.do_compile()
+
+    scope = root.get_child("__config__").scope
+    b = scope.lookup("b").get_value()
+    assert b["a"] == "a"
+    assert b["b"] == "b"
+    assert b["c"] == 3
+
+
+def test_dict_collide(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+a = "a"
+b = { "a" : a, "a" : "b", "c" : 3}
+""")
+
+    with pytest.raises(DuplicateException):
+        compiler.do_compile()
+
+
+def test_dict_attr(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+entity Foo:
+  dict bar
+  dict foo = {}
+  dict blah = {"a":"a"}
+end
+
+implement Foo using std::none
+
+a=Foo(bar={})
+b=Foo(bar={"a":z})
+c=Foo(bar={}, blah={"z":"y"})
+z=5
+""")
+
+    (_, root) = compiler.do_compile()
+
+    scope = root.get_child("__config__").scope
+
+    def map_assert(in_dict, expected):
+        for (ek, ev), (k, v) in zip(expected.items(), in_dict.items()):
+            assert ek == k
+            assert ev == v
+
+    def validate(var, bar, foo, blah):
+        e = scope.lookup(var).get_value()
+        map_assert(e.get_attribute("bar").get_value(), bar)
+        map_assert(e.get_attribute("foo").get_value(), foo)
+        map_assert(e.get_attribute("blah").get_value(), blah)
+
+    validate("a", {}, {}, {"a": "a"})
+    validate("b", {"a": 5}, {}, {"a": "a"})
+
+    validate("c", {}, {}, {"z": "y"})
+
+
+def test_dict_attr_type_error(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+entity Foo:
+  dict bar
+  dict foo = {}
+  dict blah = {"a":"a"}
+end
+
+implement Foo using std::none
+
+a=Foo(bar=b)
+b=Foo(bar={"a":"A"})
+""")
+    with pytest.raises(RuntimeException):
+        compiler.do_compile()
